dedupe cache res conversions
